+++ c/public/lib/source/uforead/uforead.c
@@ -3611,7 +3611,7 @@ int ufoEndFont(ufoCtx h) {
 }
 
 int ufoIterateGlyphs(ufoCtx h, abfGlyphCallbacks* glyph_cb) {
-    unsigned short i;
+    long i;
 
     /* Set error handler */
     DURING_EX(h->err.env)
